Handle cleaning the temporary sqlite db created when not using faitout for the tests

+++ tests/__init__.py
@@ -118,18 +118,20 @@ class Modeltests(unittest.TestCase):
     # pylint: disable=C0103
     def setUp(self):
         """ Set up the environnment, ran before every tests. """
-        dbfile = DB_PATH.split('//')[1]
-        if os.path.exists(dbfile):
-            os.unlink(dbfile)
+        if '///' in DB_PATH:
+            dbfile = DB_PATH.split('///')[1]
+            if os.path.exists(dbfile):
+                os.unlink(dbfile)
         self.session = model.create_tables(DB_PATH, debug=False)
         APP.before_request(FAS._check_session)
 
     # pylint: disable=C0103
     def tearDown(self):
         """ Remove the test.db database if there is one. """
-        dbfile = DB_PATH.split('//')[1]
-        if os.path.exists(dbfile):
-            os.unlink(dbfile)
+        if '///' in DB_PATH:
+            dbfile = DB_PATH.split('///')[1]
+            if os.path.exists(dbfile):
+                os.unlink(dbfile)
 
         self.session.rollback()
